PS: Display waiting for flags
This adds support for the recently introduced thread flags mechanism to
the `ps` module; while previously it would show a thread in state
"(null)" when it is blocked on thread_flags_wait_any or
thread_flags_wait_all, it now shoes the state "bl anyfl" or "bl allfl",
respectively. The labels are kept that short to not mess with the
fixed-width layout.